pyproject.toml
[tool.poetry]
authors = ["Naveen Thurimerla <nawinto99@gmail.com>"]
description = "e(g)it --> e(d)it, e(g)it is an app for replacing text in many files across multiple GitHub repositories."
documentation = "https://datadlog.github.io/egit/"
homepage = "https://github.com/datadlog/egit"
license = "MIT"
maintainers = ["Naveen Thurimerla <nawinto99@gmail.com>"]
name = "egit"
readme = "README.md"
repository = "https://github.com/datadlog/egit"
version = "1.1.0"

[tool.poetry.dependencies]

# Mandatory
GitPython = "^3.1.18"
PyYAML = "^5.4.1"
python = "^3.8"

# Optional
black = {version = "^21.6b0", allow-prereleases = true}
flake8 = {version = "^3.9.2", optional = true}
isort = {version = "^5.9.1", optional = true}
mkdocs = {version = "^1.2.1", optional = true}
mkdocs-autorefs = {version = "^0.2.1", optional = true}
mkdocs-include-markdown-plugin = {version = "^3.1.4", optional = true}
mkdocs-material = {version = "^7.1.9", optional = true}
mkdocs-material-extensions = {version = "^1.0.1", optional = true}
mkdocstrings = {version = "^0.15.2", optional = true}
pip = {version = "^21.1.3", optional = true}
pytest = {version = "^6.2.4", optional = true}
pytest-cov = {version = "^2.12.1", optional = true}
toml = {version = "^0.10.2", optional = true}
tox = {version = "^3.23.1", optional = true}
twine = {version = "^3.4.1", optional = true}
virtualenv = {version = "^20.4.7", optional = true}

[tool.poetry.extras]
dev = ["virtualenv", "pip", "toml", "tox", "twine"]
doc = ["mkdocs", "mkdocs-include-markdown-plugin", "mkdocs-material", "mkdocstrings", "mkdocs-material-extensions", "mkdocs-autorefs"]
test = ["pytest", "black", "isort", "flake8", "pytest-cov"]

[tool.poetry.dev-dependencies]
pylint = "^2.9.3"

[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core>=1.0.0"]

[tool.poetry.urls]
"Bug Tracker" = "https://github.com/datadlog/egit/issues"

[tool.poetry.scripts]
egit = "egit:main"
